requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-telegram-bot==13.15
pymongo
APScheduler
python-dotenv